(ciaodbg) Fix warning in conversion between bool_t and a tagged term
It was using MakeInteger, which introduced a comparison that was always
true. A generic solution consists in casting to intmatch_t, but that
still does a dummy comparison if the compiler is not smart enough.

A better solution (implemented here) use MakeSmall instead of
MakeInteger, since C booleans ({0,1}) are in the range of Ciao small
integers both for 32 and 64 bits.

Src-commit: a06ea8d3f518ef8ca1990fedbb44f85c2a6c25c7